bug hunting WIP: invalid PropTypes: books is Number
setState appears to update books ARRAY correctly:
    console.log ofBooksApp.state.book.shelf holds the NewShelf value.
      And it prints out as an array of proper shape, and data.
Yet, <ListBooks/>, and <Bookshelf/> Warns that props is getting a Number.
But if `books` is a Number, how is it able to Display them ?

Warning happens not on First Render - but only after BookShelfChanger
  initiates (and completes) BookApp's setState for book's value.
Subsequent attempts to BookshelfChanger, do show the books array === 7
	(same as books.length) -- a Number, as PropTypes indicated.
Something weird about console.log of props, or a change of its values later.