bug: WIP does not rerender after changeBookshelf
OK. Tons of code and commented out code attempts trying to get this figured out.
At this point, it *Looks Like* setState occurred, but app doesn't rerender
I'm also getting a PropTypes warning that books is a `number`, not an `array` ?!! from what I see, books is still an array of book.